Split the various profiling strategies that can be requested on the command-line from the `Profiler` implementations, so that multiple strategies can share a `Profiler` implementation.
