+++ src/main/java/org/gradle/profiler/CompositeProfiler.java
@@ -1,7 +1,5 @@
 package org.gradle.profiler;
 
-import joptsimple.OptionSet;
-
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
@@ -37,9 +35,9 @@ class CompositeProfiler extends Profiler {
     @Override
     public ProfilerController newController(final String pid, final ScenarioSettings settings) {
         List<ProfilerController> controllers = delegates.stream()
-                .map((Profiler prof) -> prof.newController(pid,
-                        settingsFor(prof, settings)))
-                .collect(Collectors.toList());
+            .map((Profiler prof) -> prof.newController(pid,
+                settingsFor(prof, settings)))
+            .collect(Collectors.toList());
         return new ProfilerController() {
             @Override
             public void startSession() throws IOException, InterruptedException {
@@ -106,7 +104,7 @@ class CompositeProfiler extends Profiler {
 
     @Override
     public GradleArgsCalculator newGradleArgsCalculator(ScenarioSettings settings) {
-        return new GradleArgsCalculator(){
+        return new GradleArgsCalculator() {
             @Override
             public void calculateGradleArgs(List<String> gradleArgs) {
                 delegates.forEach(prof -> prof.newGradleArgsCalculator(settingsFor(prof, settings)).calculateGradleArgs(gradleArgs));
@@ -116,16 +114,11 @@ class CompositeProfiler extends Profiler {
 
     @Override
     public GradleArgsCalculator newInstrumentedBuildsGradleArgsCalculator(ScenarioSettings settings) {
-        return new GradleArgsCalculator(){
+        return new GradleArgsCalculator() {
             @Override
             public void calculateGradleArgs(List<String> gradleArgs) {
                 delegates.forEach(prof -> prof.newInstrumentedBuildsGradleArgsCalculator(settingsFor(prof, settings)).calculateGradleArgs(gradleArgs));
             }
         };
-    }
-
-    @Override
-    public Profiler withConfig(OptionSet parsedOptions) {
-        return new CompositeProfiler(delegates.stream().map(profiler -> profiler.withConfig(parsedOptions)).collect(Collectors.toList()));
     }
 }
