+++ src/main/java/org/gradle/profiler/yourkit/YourKitProfiler.java
@@ -1,11 +1,7 @@
 package org.gradle.profiler.yourkit;
 
-import joptsimple.OptionParser;
-import joptsimple.OptionSet;
-import joptsimple.OptionSpecBuilder;
 import org.gradle.profiler.InstrumentingProfiler;
 import org.gradle.profiler.JvmArgsCalculator;
-import org.gradle.profiler.Profiler;
 import org.gradle.profiler.ScenarioSettings;
 
 import java.io.File;
@@ -13,14 +9,8 @@ import java.util.function.Consumer;
 
 public class YourKitProfiler extends InstrumentingProfiler {
     private final YourKitConfig yourKitConfig;
-    private OptionSpecBuilder memory;
-    private OptionSpecBuilder sampling;
 
-    public YourKitProfiler() {
-        this(null);
-    }
-
-    private YourKitProfiler(YourKitConfig yourKitConfig) {
+    YourKitProfiler(YourKitConfig yourKitConfig) {
         this.yourKitConfig = yourKitConfig;
     }
 
@@ -49,24 +39,5 @@ public class YourKitProfiler extends InstrumentingProfiler {
     @Override
     protected JvmArgsCalculator jvmArgsWithInstrumentation(ScenarioSettings settings, boolean startRecordingOnProcessStart, boolean captureSnapshotOnProcessExit) {
         return new YourKitJvmArgsCalculator(settings, yourKitConfig, startRecordingOnProcessStart, captureSnapshotOnProcessExit);
-    }
-
-    @Override
-    public void addOptions(OptionParser parser) {
-        memory = parser.accepts("yourkit-memory", "Perform memory profiling instead of CPU profiling");
-        sampling = parser.accepts("yourkit-sampling", "Use sampling instead of tracing for CPU profiling");
-    }
-
-    @Override
-    public Profiler withConfig(OptionSet parsedOptions) {
-        return new YourKitProfiler(newConfigObject(parsedOptions));
-    }
-
-    private YourKitConfig newConfigObject(OptionSet parsedOptions) {
-        YourKitConfig yourKitConfig = new YourKitConfig(parsedOptions.has(memory), parsedOptions.has(sampling));
-        if (yourKitConfig.isMemorySnapshot() && yourKitConfig.isUseSampling()) {
-            throw new IllegalArgumentException("Cannot use memory profiling and CPU sampling at the same time.");
-        }
-        return yourKitConfig;
     }
 }
